+++ example/tailor/packages/display/vite.config.ts
@@ -23,7 +23,7 @@ export default defineConfig({
       // into your library
       external: ['vue'],
       output: {
-        intro: 'import "./style.css";',
+        intro: 'import "./index.css";',
         // Provide global variables to use in the UMD build
         // for externalized deps
         globals: {