+++ packages/runtime/tce-display/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @tailor-cms/tce-display-runtime
 
+## 0.6.9
+
+### Patch Changes
+
+- Updated dependencies.
+
 ## 0.6.8
 
 ### Patch Changes