+++ packages/runtime/tce-edit/CHANGELOG.md
@@ -1,6 +1,13 @@
 # @tailor-cms/tce-edit-runtime
 
+## 1.0.1
+
+### Patch Changes
+
+- Updated dependencies.
+
 ## 1.0.0
 
 ### Major Changes
+
 - Package is now Vue 3 based, replacing tce-edit-next.