+++ packages/runtime/tce-server/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @tailor-cms/tce-server-runtime
 
+## 0.5.6
+
+### Patch Changes
+
+- Updated dependencies.
+
 ## 0.5.5
 
 ### Patch Changes